requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1615404881292/work
argh @ file:///home/conda/feedstock_root/build_artifacts/argh_1595627874344/work
blinker==1.4
brotlipy==0.7.0
cachetools==2.0.0
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1613246439577/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093477613/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1621503698523/work
cloudpickle==1.3.0
configparser @ file:///home/conda/feedstock_root/build_artifacts/configparser_1614550305463/work
continuum==1.0.2
crc32c==2.2.post0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1616851478720/work
cycler==0.10.0
cytoolz==0.11.0
dask==1.1.4
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==4.4.2
denoising-diffusion-pytorch==0.6.9
docker-pycreds==0.4.0
einops==0.3.2
future==0.18.2
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1616781027402/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1620970877194/work
google-auth==1.35.0
google-auth-oauthlib==0.4.1
gql==0.1.0
graphql-core @ file:///home/conda/feedstock_root/build_artifacts/graphql-core_1621603344031/work
grpcio @ file:///home/conda/feedstock_root/build_artifacts/grpcio_1604365516153/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1622488238537/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282412546/work
linformer==0.2.1
lmdb==1.2.1
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1614595805172/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
networkx==2.2
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1604945999799/work
nvidia-ml-py3==7.352.0
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1622383516566/work
olefile==0.46
pathtools==0.1.2
Pillow @ file:///tmp/build/80754af9/pillow_1617386154532/work
promise @ file:///home/conda/feedstock_root/build_artifacts/promise_1610108967426/work
protobuf==3.14.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1603570993785/work
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1619620644068/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291458349/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658315997/work
PyYAML==5.3.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1595492159598/work
rsa==3.4.2
scikit-image==0.17.2
scipy==1.4.1
sentry-sdk @ file:///home/conda/feedstock_root/build_artifacts/sentry-sdk_1620758041679/work
shortuuid @ file:///home/conda/feedstock_root/build_artifacts/shortuuid_1610439219919/work
six @ file:///tmp/build/80754af9/six_1605205335545/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1611376390914/work
subprocess32 @ file:///tmp/build/80754af9/subprocess32_1602000156065/work
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1610699261066/work/tensorboard-2.4.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1611075653546/work/tensorboard_plugin_wit-1.8.0-py3-none-any.whl
tensorboardX==2.4
tfrecord==1.14.1
tifffile==2020.10.1
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
torch==1.6.0
torchaudio==0.8.0a0+e4e171a
torchvision==0.7.0
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1621890532941/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1622056799390/work
vit-pytorch==0.22.0
wandb @ file:///home/conda/feedstock_root/build_artifacts/wandb_1622207453758/work
watchdog @ file:///home/conda/feedstock_root/build_artifacts/watchdog_1605984115630/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1621518206714/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
